refactor(overview): changed mock file location
